initial commit
totally rework auton selection, no touch screen now due to a crack on the new brain. also i made a new pid that allows the bot to move to a destination, and slow down as it is reached. this is great for auton. also a bunch of optimizations and tweaks boiii