prototype position tracking code
based off of the e-bots pilons' (5225's) "Introduction to Position Tracking." the code is located in odom.h. Without testing, i have no clue if it works yet. I'm very thankful for 5225's document, and I hope I interpreted it right. Keep in mind that the code was written in just a few hours and there is a lot of rewriting/optimization to do.