final changes to xdrive code for the season
as clunky and unfinished as it is, this code project will no longer be used and will be pushed to a new branch. we have redesigned the robot and will be using an entirely new project, this time without xdrive or field centric. We will still use the same odometry code, however. (adjusted for the new rotation sensor)