finish all 4 autos, make log volt drive work
we got a whole lotta functions for different pid values with different centers of gravity and distances and shit but yeah all 4 autos work now, and I fixed the log volt drive. the problem was that the 12/100 was casting to an integer which made it 0, so the drive value was always 0. ready for daniel hand!