final commit, goodbye
the last version of our code ill ever touch. went to worlds with it, made it to round of 16 in our division. Very proud of our other teams, 4478D won the division inspire award, while 4478E went on to the semifinals in the dome, becoming one of the top 8 teams in the world. Congrats to everyone, it was a wonderful few years ive had.